Fix compilation error

+++ core/src/server/grpc_impl/GrpcServer.cpp
@@ -50,7 +50,7 @@ class NoReusePortOption : public ::grpc::ServerBuilderOption {
     void
     UpdateArguments(::grpc::ChannelArguments* args) override {
         args->SetInt(GRPC_ARG_ALLOW_REUSEPORT, 0);
-        unsigned int grpc_concurrency = 4 * std::thread::hardware_concurrency();
+        long grpc_concurrency = 4 * std::thread::hardware_concurrency();
         grpc_concurrency = std::max(32, grpc_concurrency);
         grpc_concurrency = std::min(256, grpc_concurrency);
         args->SetInt(GRPC_ARG_MAX_CONCURRENT_STREAMS, grpc_concurrency);
